added deleteaccount and admin terminal security

+++ app.py
@@ -156,24 +156,51 @@ def logout():
 @app.route("/admin", methods=["GET", "POST"])
 @login_required
 def admin():
-    return render_template("admin.html")
+    user = User.query.filter_by(username="admin0909").first()
+    if current_user.username == "admin0909" and bcrypt.check_password_hash(user.password, "admin0909"):
+        return render_template("admin.html")
+    else:
+        return redirect(url_for("board"))
 
 @app.route("/delete", methods=["GET", "POST"])
-@login_required
 def delete():
-    db.session.query(Message).delete()
-    db.session.commit()
-    return redirect(url_for("admin"))
+    user = User.query.filter_by(username="admin0909").first()
+    if current_user.username == "admin0909" and bcrypt.check_password_hash(user.password, "admin0909"):
+            db.session.query(Message).delete()
+            db.session.commit()
+            return redirect(url_for("board"))
+    else:
+        return redirect(url_for("board"))
+
+
+
 
 
 @app.route("/deleteid", methods=["POST","GET"])
-@login_required
 def deleteid():
-    data = request.form.get('idform', 0)
-    int(data)
-    Message.query.filter(Message.id == data).delete()
-    db.session.commit()
-    return redirect(url_for("admin"))
+    user = User.query.filter_by(username="admin0909").first()
+    if current_user.username == "admin0909" and bcrypt.check_password_hash(user.password, "admin0909"):
+        data = request.form.get('idform', 0)
+        int(data)
+        Message.query.filter(Message.id == data).delete()
+        db.session.commit()
+        return redirect(url_for("board"))
+    else:
+        return redirect(url_for("board"))
+
+
+@app.route("/deleteaccount", methods=["POST","GET"])
+def deleteaccount():
+    user = User.query.filter_by(username="admin0909").first()
+    if current_user.username == "admin0909" and bcrypt.check_password_hash(user.password, "admin0909"):
+            account = request.form.get("accountid",0)
+            int(account)
+            User.query.filter(User.id == account).delete()
+            db.session.commit()
+            return redirect(url_for("board"))
+    else:
+        return redirect(url_for("board"))
+
 
 if __name__ == "__main__":
     app.run(debug=True)
